releases/V2_Stable_Release_20260320_100509/requirements.txt
# Requirements for Synthetic Maternity Research Pipeline
# Auto-generated by release_packager.py
#

anthropic
datasets
google-generativeai
matplotlib
numpy
openai
pandas
python-dotenv
requests
scipy
